+++ fs/symlink-dependency/test/safeJoinModulesDir.test.ts
